Moving Triggers to their own class
